--- src/compas/robots/base_artist/__init__.py
@@ -1,7 +1,0 @@
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-
-from ._artist import *  # noqa: F401 F403
-
-__all__ = [name for name in dir() if not name.startswith('_')]
